# generated-code/src/gen/java/io/swagger/model/TestScript.java
/*
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.model.Link;
import io.swagger.v3.oas.annotations.media.Schema;
import javax.validation.constraints.*;
import org.eclipse.lyo.oslc4j.core.annotation.*;
import org.eclipse.lyo.oslc4j.core.model.AbstractResource;
/**
 * TestScript
 */
@OslcResourceShape(title = "Test Script Resource Shape", describes = "http://open-services.net/ns/qm#TestScript")
@OslcNamespace("http://open-services.net/ns/qm#")
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen", date = "2020-05-01T13:27:13.952-05:00[America/Mexico_City]")public class TestScript  extends AbstractResource {
  @JsonProperty("title")
  private String title = null;

  @JsonProperty("validatesRequirement")
  private Link validatesRequirement = null;

  @JsonProperty("description")
  private String description = null;

  public TestScript title(String title) {
    this.title = title;
    return this;
  }

  /**
   * Get title
   * @return title
   **/
  @OslcPropertyDefinition("http://purl.org/dc/terms/title")
  @JsonProperty("title")
  @Schema(description = "")
  public String getTitle() {
    return title;
  }

  public void setTitle(String title) {
    this.title = title;
  }

  public TestScript validatesRequirement(Link validatesRequirement) {
    this.validatesRequirement = validatesRequirement;
    return this;
  }

  /**
   * Get validatesRequirement
   * @return validatesRequirement
   **/
  @JsonProperty("validatesRequirement")
  @Schema(description = "")
  public Link getValidatesRequirement() {
    return validatesRequirement;
  }

  public void setValidatesRequirement(Link validatesRequirement) {
    this.validatesRequirement = validatesRequirement;
  }

  public TestScript description(String description) {
    this.description = description;
    return this;
  }

  /**
   * Get description
   * @return description
   **/
  @OslcPropertyDefinition("http://purl.org/dc/terms/description")
  @JsonProperty("description")
  @Schema(description = "")
  public String getDescription() {
    return description;
  }

  public void setDescription(String description) {
    this.description = description;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    TestScript testScript = (TestScript) o;
    return Objects.equals(this.title, testScript.title) &&
        Objects.equals(this.validatesRequirement, testScript.validatesRequirement) &&
        Objects.equals(this.description, testScript.description);
  }

  @Override
  public int hashCode() {
    return Objects.hash(title, validatesRequirement, description);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class TestScript {\n");
    
    sb.append("    title: ").append(toIndentedString(title)).append("\n");
    sb.append("    validatesRequirement: ").append(toIndentedString(validatesRequirement)).append("\n");
    sb.append("    description: ").append(toIndentedString(description)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}